requirements.txt
streamlit
gitpython
pillow 